# cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample YourVibes backend server",
        "title": "API Documentation YourVibes backend",
        "termsOfService": "https://github.com/poin4003/yourVibes_GoApi",
        "contact": {
            "name": "TEAM HKTP",
            "url": "https://github.com/poin4003/yourVibes_GoApi",
            "email": "pchuy4003@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "yourvibesapi.duckdns.org:8080",
    "basePath": "/v1/2024",
    "paths": {
...
        "/comments/like_comment/{comment_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve multiple user is like comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "like_comment"
                ],
                "summary": "Get User like comments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "comment ID to get user like comment",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit of users per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "When user like comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "like_comment"
                ],
                "summary": "Like comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "comment ID to create like comment",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/users/auth_google/": {
            "post": {
                "description": "When user need google login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_auth"
                ],
                "summary": "User auth google",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AuthGoogleRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/change_password/": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "When user need to change password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_auth"
                ],
                "summary": "User change password",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/forgot_user_password/": {
            "post": {
                "description": "When user forgot and change password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_auth"
                ],
                "summary": "User forgot password",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ForgotUserPasswordRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/users/get_otp_forgot_user_password/": {
            "post": {
                "description": "Before forgot password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_auth"
                ],
                "summary": "User get otp forgot user password",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GetOtpForgotUserPasswordRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/login/": {
            "post": {
                "description": "When user login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/users/register/": {
            "post": {
                "description": "When user registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_auth"
                ],
                "summary": "User Registration",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RegisterRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/users/verifyemail/": {
            "post": {
                "description": "Before user registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_auth"
                ],
                "summary": "User verify email",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.VerifyEmailRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "consts.Platform": {
            "type": "string",
            "enum": [
                "web",
                "android",
                "ios"
            ],
            "x-enum-varnames": [
                "WEB",
                "ANDROID",
                "IOS"
            ]
        },
...
        "request.AuthGoogleRequest": {
            "type": "object",
            "properties": {
                "authorization_code": {
                    "type": "string"
                },
                "platform": {
                    "$ref": "#/definitions/consts.Platform"
                }
            }
        },
...
        "request.ChangePasswordRequest": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
        "request.ForgotUserPasswordRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                },
                "otp": {
                    "type": "string"
                }
            }
        },
        "request.GetOtpForgotUserPasswordRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "request.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "request.RegisterRequest": {
            "type": "object",
            "required": [
                "birthday",
                "email",
                "family_name",
                "name",
                "otp",
                "password",
                "phone_number"
            ],
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "family_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "otp": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
        "request.VerifyEmailRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Token without 'Bearer ' prefix",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}